oidc: harden reloadable confirmation flow
Updates #3365
