+++ hscontrol/handlers.go
@@ -40,18 +40,23 @@ func httpError(w http.ResponseWriter, err error) {
 // an actionable message derived from the HTTP status code.
 func httpUserError(w http.ResponseWriter, err error) {
 	code := http.StatusInternalServerError
+	userMsg := ""
 
 	if herr, ok := errors.AsType[HTTPError](err); ok {
 		if herr.Code != 0 {
 			code = herr.Code
 		}
 
+		userMsg = herr.UserMsg
+
 		log.Error().Err(herr.Err).Int("code", code).Msgf("user msg: %s", herr.Msg)
 	} else {
 		log.Error().Err(err).Int("code", code).Msg("http internal server error")
 	}
 
-	userMsg := userMessageForStatusCode(code)
+	if userMsg == "" {
+		userMsg = userMessageForStatusCode(code)
+	}
 
 	w.Header().Set("Content-Type", "text/html; charset=utf-8")
 	w.WriteHeader(code)
@@ -73,10 +78,7 @@ func userMessageForStatusCode(code int) string {
 	case code == http.StatusUnauthorized || code == http.StatusForbidden:
 		return "You are not authorized. Please contact your administrator."
 	case code == http.StatusGone:
-		// Overwhelmingly a reload or a back button on a link the user
-		// already used, not a failure, so lead with that.
-		return "This link has already been used or has expired. " +
-			"If your device is connected you are done; otherwise start the login again."
+		return "Your session has expired. Please try again."
 	case code >= 400 && code < 500:
 		return "The request could not be processed. Please try again."
 	default:
@@ -86,9 +88,10 @@ func userMessageForStatusCode(code int) string {
 
 // HTTPError represents an error that is surfaced to the user via web.
 type HTTPError struct {
-	Code int    // HTTP response code to send to client; 0 means 500
-	Msg  string // Response body to send to client
-	Err  error  // Detailed error to log on the server
+	Code    int    // HTTP response code to send to client; 0 means 500
+	Msg     string // Response body to send to non-browser clients
+	Err     error  // Detailed error to log on the server
+	UserMsg string // Optional safe message for browser-facing error pages
 }
 
 func (e HTTPError) Error() string { return fmt.Sprintf("http error[%d]: %s, %s", e.Code, e.Msg, e.Err) }
@@ -97,6 +100,10 @@ func (e HTTPError) Unwrap() error { return e.Err }
 // NewHTTPError returns an HTTPError containing the given information.
 func NewHTTPError(code int, msg string, err error) HTTPError {
 	return HTTPError{Code: code, Msg: msg, Err: err}
+}
+
+func newHTTPUserError(code int, msg, userMsg string, err error) HTTPError {
+	return HTTPError{Code: code, Msg: msg, Err: err, UserMsg: userMsg}
 }
 
 var errMethodNotAllowed = NewHTTPError(http.StatusMethodNotAllowed, "method not allowed", nil)
