+++ hscontrol/handlers_test.go
@@ -202,8 +202,20 @@ func TestHttpUserError(t *testing.T) {
 			name:           "gone_renders_session_expired",
 			err:            NewHTTPError(http.StatusGone, "login session expired, try again", nil),
 			wantCode:       http.StatusGone,
-			wantContains:   "This link has already been used or has expired.",
+			wantContains:   "Your session has expired. Please try again.",
 			wantNotContain: "login session expired",
+		},
+		{
+			name: "gone_with_user_message_renders_specific_guidance",
+			err: newHTTPUserError(
+				http.StatusGone,
+				"registration link already used or expired",
+				"This link has already been used or has expired.",
+				nil,
+			),
+			wantCode:       http.StatusGone,
+			wantContains:   "This link has already been used or has expired.",
+			wantNotContain: "registration link already used or expired",
 		},
 		{
 			name:           "bad_request_renders_generic_retry",
